api/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
passlib[argon2]
argon2-cffi
python-jose[cryptography]
email-validator
